Functions are Hard
Was trying to learn/understand passing arguments by reference. It's
slow, but working. I'm thinking that I will have a function,
playerAnswer, that gets a user input one at a time and then passes that
back to the loop to be checked by answerSheet.

--or pass the answer in answerSheet and obtain/check against the
playerInput, and then give back a right/wrong.

We'll see.